refactor: Line 메소드 중 Sections 도메인으로 리팩터링
